Build a Package from the code
