+++ Main/DB/__init__.py
@@ -1,0 +1,2 @@
+from .DatabaseHandler import *
+from .SQLQueries import *
