+++ Main/HTML/__init__.py
@@ -1,0 +1,4 @@
+from .HTML import *
+from .html_files import *
+from .HTML_TEMPLATES import *
+from .JS_TEMPLATES import *
