--- Tests_Examples/a_test.py
@@ -1,11 +1,0 @@
-from code.code2 import *
-
-
-def test1():
-    assert(func1(5) == 6)
-
-
-def test2():
-    assert(func2([3, 3], 1) == 3)
-
-
