--- Tests_Examples/b_test.py
@@ -1,9 +1,0 @@
-from code2.code2 import *
-
-
-def test1():
-    assert (func4(5) == 6)
-
-
-def test2():
-    assert (func5([3, 3], 1) == 3)
