--- code/code2.py
@@ -1,13 +1,0 @@
-def func1(x):
-    return x + 1
-
-
-def func2(arr, i):
-    return arr[i]
-
-
-def func3(isTrue):
-    if isTrue:
-        return "True!"
-    else:
-        return "False!"
