--- code/code4.py
@@ -1,13 +1,0 @@
-def func4(x):
-    return x + 1
-
-
-def func5(arr, i):
-    return arr[i]
-
-
-def func6(isTrue):
-    if isTrue:
-        return "True!"
-    else:
-        return "False!"
